requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.1
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
deepface==0.0.93
tensorflow==2.18.0
